+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Client mode:
 * PR #237: Better generation for parameters (thanks to klafbang)
     * Setting a bind variable to null would erroneously remove the relevant argument in the query
     * Default value in GraphQL request parameters was not managed
+* Correction of a NullPointerException when a query returns an interface containing fields with parameters
 
 
 ## 3.0.1